# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Book Search",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/organization/destinations/{root_key}": {
      "put": {
        "tags": [
          "organization"
        ],
        "summary": "Save Destination",
        "operationId": "save_destination_api_organization_destinations__root_key__put",
        "parameters": [
          {
            "name": "root_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Root Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DestinationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organization/destinations/{destination_id}/probe": {
      "post": {
        "tags": [
          "organization"
        ],
        "summary": "Probe Destination",
        "operationId": "probe_destination_api_organization_destinations__destination_id__probe_post",
        "parameters": [
          {
            "name": "destination_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Destination Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProbeInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organization/destinations/{destination_id}/automatic-import": {
      "get": {
        "tags": [
          "organization"
        ],
        "summary": "Detail",
        "operationId": "detail_api_organization_destinations__destination_id__automatic_import_get",
        "parameters": [
          {
            "name": "destination_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Destination Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "organization"
        ],
        "summary": "Save",
        "operationId": "save_api_organization_destinations__destination_id__automatic_import_put",
        "parameters": [
          {
            "name": "destination_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Destination Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DestinationInput": {
        "properties": {
          "library_id": {
            "type": "string",
            "format": "uuid",
            "title": "Library Id"
          },
          "medium": {
            "type": "string",
            "enum": [
              "ebook",
              "audio"
            ],
            "title": "Medium"
          },
          "backend_path": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 2,
            "title": "Backend Path"
          },
          "mode": {
            "type": "string",
            "enum": [
              "hardlink",
              "copy"
            ],
            "title": "Mode",
            "default": "hardlink"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "expected_revision": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-f0-9]{64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "library_id",
          "medium",
          "backend_path"
        ],
        "title": "DestinationInput"
      },
      "DestinationView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "root_key": {
            "type": "string",
            "title": "Root Key"
          },
          "library_id": {
            "type": "string",
            "format": "uuid",
            "title": "Library Id"
          },
          "medium": {
            "type": "string",
            "title": "Medium"
          },
          "backend_path": {
            "type": "string",
            "title": "Backend Path"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "revision": {
            "type": "string",
            "title": "Revision"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "probe": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probe"
          },
          "publication_available": {
            "type": "boolean",
            "title": "Publication Available",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "root_key",
          "library_id",
          "medium",
          "backend_path",
          "mode",
          "enabled",
          "revision",
          "configured",
          "probe"
        ],
        "title": "DestinationView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperationView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "message",
          "created_at",
          "updated_at"
        ],
        "title": "OperationView"
      },
...
      "PolicyInput": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "expected_generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Generation"
          },
          "destination_revision": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Destination Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled",
          "expected_generation",
          "destination_revision"
        ],
        "title": "PolicyInput"
      },
      "PolicyView": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "generation": {
            "type": "integer",
            "title": "Generation"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "can_enable": {
            "type": "boolean",
            "title": "Can Enable"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled",
          "generation",
          "ready",
          "can_enable",
          "message"
        ],
        "title": "PolicyView"
      },
...
      "ProbeInput": {
        "properties": {
          "plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Plan Id"
          },
          "expected_revision": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_id",
          "expected_revision"
        ],
        "title": "ProbeInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}